smartpos/backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
typing-extensions>=4.7.1
pydantic>=2.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0
email-validator>=2.0.0
python-dateutil>=2.8.2
Pillow>=10.0.0
aiofiles>=23.0.0
